+++ frontend/src/app/(admin)/faqs-editor/page.tsx
@@ -23,8 +23,8 @@ import { type FaqItem, type FaqPage, getFaq, putFaqs } from "@/api/faq";
 import { ConfirmationNotification } from "@/components/admin-portal/ConfirmationNotification";
 import { DraggableSortablePill } from "@/components/admin-portal/DraggableSortablePill";
 import { FaqCard } from "@/components/admin-portal/FaqCard";
-import { PreviewMode } from "@/components/admin-portal/PreviewMode";
-import { PreviewNavBar } from "@/components/admin-portal/PreviewNavBar";
+import { PreviewMode } from "@/components/admin-portal/preview-components/PreviewMode";
+import { PreviewNavBar } from "@/components/admin-portal/preview-components/PreviewNavBar";
 import { PublishButton } from "@/components/admin-portal/PublishButton";
 import { RevertButton } from "@/components/admin-portal/RevertButton";
 import { ConfirmationDialog } from "@/components/ConfirmationDialog";
