+++ frontend/src/app/(admin)/timeline-editor/page.tsx
@@ -7,8 +7,8 @@ import type { StorageReference } from "firebase/storage";
 
 import { getTimelines, type TimelineItem, updateTimeline } from "@/api/timeline";
 import { HeaderSection } from "@/components/admin-portal/HeaderSection";
-import { PreviewMode } from "@/components/admin-portal/PreviewMode";
-import { PreviewNavBar } from "@/components/admin-portal/PreviewNavBar";
+import { PreviewMode } from "@/components/admin-portal/preview-components/PreviewMode";
+import { PreviewNavBar } from "@/components/admin-portal/preview-components/PreviewNavBar";
 import { PublishButton } from "@/components/admin-portal/PublishButton";
 import { RevertButton } from "@/components/admin-portal/RevertButton";
 import { TimelineCard } from "@/components/admin-portal/TimelineCard";
